third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/embed-tag-code-attribute-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3Cembed%20code=data:text/html%3bbase64,PHNjcmlwdD5hbGVydCgxKTwvc2NyaXB0Pg==%3E%3C/embed%3E' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
